Adds ability to create a new library dynamically
Logged in users can now create a new collections via a dialog box. This creates a new table in the MySQL instance with the proper parameters, adds this table name to a table that tracks users and table names and reflects this new table in the UI immediately. Different users can only access their own tables.