Adds logout and change library functionality
The changes in this commit allow the user to logout from both the collection manager window and the record manager window. When the user logs out, they are taken back to the main login screen to enter a new username/password, or register a new account. 

This commit also lets users change record collections from the record manager screen by clicking a button 'Change Library'. This brings users back to the collection manager window so they can open, create or delete a new library. This is all done through the use of "QDialog::done(int)" method. Depending on the user action, a different int value is returned by the done() method to main.cpp. Goto statements control the program flow to allow windows to be re-opened.

Proper handling of clearing member variables has also been added, so that when windows are closed, they are fully cleaned out so they can be re-constructed and re-initialized from scratch.